Fixed some logic and extracted progress bar and next button code to new components
