+++ code/src/components/ProgressBar.jsx
@@ -1,0 +1,12 @@
+import React from "react"
+
+const ProgressBar = ({ step }) => {
+
+  return (
+      <div className="progress">
+        <div className={`progress-value page-${step}`}></div>
+      </div>
+  )
+}
+
+export default ProgressBar
