+++ code/src/components/QuestionRadioVehicle.jsx
@@ -1,6 +1,6 @@
 import React from "react"
 
-import { RandomWeather } from '../RandomChoices'
+import { RandomWeather } from "../RandomChoices"
 
 const vehicleTypes = ["bicycle", "e-scooter", "motorcycle"]
 
@@ -9,25 +9,23 @@ const QuestionRadioVehicle = (props) => {
 
   return (
     <>
-      <div className="progress">
-        <div className="progress-value question4"></div>
-      </div>
-      <form onSubmit={() => onStepChange(1)}>
-      <h1>What would you rather ride on your way to work under this {weather ? weather : RandomWeather()} sky?</h1>
-        {vehicleTypes.map((type) => (
-          <label key={type}>
-            <input
-              type="radio"
-              value={type}
-              onChange={onVehicleChange}
-              checked={vehicle === type}
-            />
-            {type}
-          </label>
-        ))}
-        <button type='submit'>Next question</button>
-        <button type="button" onClick={() => onStepChange(-1)}>Go back</button>
-      </form>
+      <h1>
+        What would you rather ride on your way to work under this{" "}
+        {weather ? weather : RandomWeather()} sky?
+      </h1>
+      {vehicleTypes.map((type) => (
+        <label key={type}>
+          <input
+            type="radio"
+            value={type}
+            onChange={onVehicleChange}
+            checked={vehicle === type}
+          />
+          {type}
+        </label>
+      ))}
+      {/* <button onClick={() => onStepChange(1)}>Next</button> */}
+      <button onClick={() => onStepChange(-1)}>Back</button>
     </>
   )
 }
