+++ code/src/components/QuestionRangeSliderSpeed.jsx
@@ -7,10 +7,6 @@ const QuestionRangeSliderSpeed = (props) => {
 
   return (
     <>
-      <div className="progress">
-        <div className="progress-value question5"></div>
-      </div>
-      <form onSubmit={() => onStepChange(1)}>
         <label>
         <h1>Let’s be honest, would you be more the “slowing the traffic” or “scaring the old ladies” type on your {vehicle ? vehicle : RandomVehicle()}?</h1>
           <input
@@ -21,9 +17,8 @@ const QuestionRangeSliderSpeed = (props) => {
             onChange={onSpeedChange}
           />
         </label>
-        <button type='submit'>Next question</button>
-        <button type="button" onClick={() => onStepChange(-1)}>Go back</button>
-      </form>
+        {/* <button onClick={() => onStepChange(1)}>Next</button> */}
+        <button onClick={() => onStepChange(-1)}>Back</button>
     </>
   )
 }
