+++ code/src/components/QuestionSelectMenuSnack.jsx
@@ -7,10 +7,6 @@ const QuestionSelectMenuSnack = (props) => {
 
   return (
     <>
-      <div className="progress">
-        <div className="progress-value question6"></div>
-      </div>
-      <form onSubmit={() => onStepChange(1)}>
       <label htmlFor="snack"><h1>I offer you a snack, {userName ? userName : RandomUserName()}. What do you choose between these options?</h1></label>
         <select onChange={onSnackChange} id="gender" value={snack}>
           <option value="" disabled>--Please select an option--</option>
@@ -18,9 +14,8 @@ const QuestionSelectMenuSnack = (props) => {
           <option value="candies">candies</option>
           <option value="nuts">nuts</option>
         </select>
-        <button type='submit'>Next question</button>
-        <button type="button" onClick={() => onStepChange(-1)}>Go back</button>
-      </form>
+        {/* <button onClick={() => onStepChange(1)}>Next</button> */}
+        <button onClick={() => onStepChange(-1)}>Back</button>
     </>
   )
 }
