+++ code/src/components/QuestionSelectMenuSports.jsx
@@ -7,10 +7,6 @@ const QuestionSelectMenuSports = (props) => {
 
   return (
     <>
-      <div className="progress">
-        <div className="progress-value question11"></div>
-      </div>
-      <form onSubmit={() => onStepChange(1)}>
       <label htmlFor="sport"><h1>Maybe you would like to watch some game with {friendName ? friendName : RandomFriendName()}: what would be the sport then?</h1></label>
         <select onChange={onSportChange} id="sport" value={sport}>
           <option value="" disabled>--Please select an option--</option>
@@ -20,9 +16,8 @@ const QuestionSelectMenuSports = (props) => {
           <option value="floorball">floorball</option>
           <option value="basketball">basketball</option>
         </select>
-        <button type='submit'>Next question</button>
-        <button type="button" onClick={() => onStepChange(-1)}>Go back</button>
-      </form>
+        {/* <button onClick={() => onStepChange(1)}>Next</button> */}
+        <button onClick={() => onStepChange(-1)}>Back</button>
     </>
   )
 }
